moved run and visualize to utils
